+++ src/Shrun/Configuration/Args/Parsing/ConsoleLogging.hs
@@ -95,9 +95,9 @@ lineTruncParser =
       mconcat
         [ "Non-negative integer that limits the length of console logs. Can ",
           "also be the string literal 'detect', to detect the terminal size ",
-          "automatically. Defaults to no truncation. Note that \"log prefixes\" ",
-          "(e.g. labels like [Success], timestamps) are counted towards the ",
-          "total length but are never truncated."
+          "automatically. Defaults to 'detect' if --console-log-command is on. ",
+          "Note that \"log prefixes\" (e.g. labels like [Success], timestamps) ",
+          "are counted towards the total length but are never truncated."
         ]
 
 stripControlParser :: Parser (Maybe ConsoleLogStripControl)